Moves service order to OrderController
