+++ packages/api/app/Controllers/Http/OrderController.js
@@ -2,7 +2,16 @@ const Technology = use('App/Models/Technology');
 const TechnologyOrder = use('App/Models/TechnologyOrder');
 const ServiceOrder = use('App/Models/ServiceOrder');
 const Permission = use('App/Models/Permission');
-const { orderStatuses, errorPayload, errors, getTransaction, permissions } = require('../../Utils');
+const ServiceOrderReview = use('App/Models/ServiceOrderReview');
+const User = use('App/Models/User');
+const {
+	orderStatuses,
+	serviceOrderStatuses,
+	errorPayload,
+	errors,
+	getTransaction,
+	permissions,
+} = require('../../Utils');
 
 const {
 	LIST_TECHNOLOGIES_ORDERS: { permission: listTechnologiesOrdersPermission },
@@ -40,6 +49,30 @@ class OrderController {
 		}
 
 		return technologyOrderQuery.withFilters(request).withParams(request, { filterById: false });
+	}
+
+	async showServiceOrders({ auth, request }) {
+		const serviceResponsible = await auth.getUser();
+		const serviceOrders = ServiceOrder.query()
+			.whereHas('service', (builder) => {
+				builder.where({ user_id: serviceResponsible.id });
+			})
+			.with('service')
+			.withParams(request);
+		return serviceOrders;
+	}
+
+	async showServiceOrderReviews({ auth, request }) {
+		const serviceResponsible = await auth.getUser();
+		const serviceOrderReviews = ServiceOrderReview.query()
+			.whereHas('serviceOrder', (builder) => {
+				builder.whereHas('service', (builder2) => {
+					builder2.where({ user_id: serviceResponsible.id });
+				});
+			})
+			.with('serviceOrder.service')
+			.withParams(request);
+		return serviceOrderReviews;
 	}
 
 	/** Lists orders in buyer and seller view */
@@ -139,6 +172,28 @@ class OrderController {
 		return order;
 	}
 
+	async updateServiceOrder({ params, request }) {
+		const { quantity, comment } = request.all();
+		const serviceOrder = await ServiceOrder.findOrFail(params.id);
+		serviceOrder.merge({ quantity, comment });
+		await serviceOrder.save();
+		return serviceOrder;
+	}
+
+	async updateServiceOrderReview({ params, request }) {
+		const data = request.only(['content', 'rating', 'positive', 'negative']);
+		const review = {
+			content: data.content,
+			rating: data.rating,
+			positive: JSON.stringify(data.positive),
+			negative: JSON.stringify(data.negative),
+		};
+		const updatedServiceOrderReview = await ServiceOrderReview.findOrFail(params.id);
+		updatedServiceOrderReview.merge(review);
+		await updatedServiceOrderReview.save();
+		return updatedServiceOrderReview.toJSON();
+	}
+
 	sendEmailToResearcher(researcher, technology, antl) {
 		const mailData = {
 			email: researcher.email,
@@ -150,7 +205,7 @@ class OrderController {
 		Bull.add(SendMailJob.key, mailData, { attempts: 3 });
 	}
 
-	async store({ auth, params, request }) {
+	async storeTechnologyOrder({ auth, params, request }) {
 		const technology = await Technology.findOrFail(params.id);
 		let technologyOrder;
 		let trx;
@@ -181,10 +236,85 @@ class OrderController {
 	}
 
 	/**
+	 * Send emails to the users responsible for services.
+	 *
+	 * @param {Array} serviceOrders Service Orders
+	 * @param {Function} antl Function to translate the messages
+	 */
+	async sendEmailsToResponsibles(serviceOrders, antl) {
+		await Promise.all(
+			serviceOrders.map(async (serviceOrder) => {
+				await serviceOrder.load('service');
+				const responsible = await User.findOrFail(serviceOrder.toJSON().service.user_id);
+				const mailData = {
+					email: responsible.email,
+					subject: antl('message.service.serviceRequested'),
+					template: 'emails.service-requested',
+					responsible,
+					serviceOrder,
+				};
+				Bull.add(SendMailJob.key, mailData, { attempts: 3 });
+			}),
+		);
+	}
+
+	async storeServiceOrder({ auth, request }) {
+		const { services, comment } = request.all();
+		const user = await auth.getUser();
+		const servicesList = services.map((service) => ({
+			service_id: service.service_id,
+			quantity: service.quantity,
+			user_id: user.id,
+			status: serviceOrderStatuses.REQUESTED,
+			comment,
+		}));
+		const serviceOrders = await user.serviceOrders().createMany(servicesList);
+		await this.sendEmailsToResponsibles(serviceOrders, request.antl);
+		return serviceOrders;
+	}
+
+	async storeServiceOrderReview({ request, params, auth }) {
+		const data = request.only(['content', 'rating', 'positive', 'negative']);
+
+		const review = {
+			content: data.content,
+			rating: data.rating,
+			positive: JSON.stringify(data.positive),
+			negative: JSON.stringify(data.negative),
+		};
+
+		const [serviceOrder, user] = await Promise.all([
+			ServiceOrder.findOrFail(params.id),
+			auth.getUser(),
+		]);
+
+		let serviceOrderReview;
+		let trx;
+
+		try {
+			const { init, commit } = getTransaction();
+			trx = await init();
+
+			serviceOrderReview = await ServiceOrderReview.create(review, trx);
+			await Promise.all([
+				serviceOrderReview.serviceOrder().associate(serviceOrder, trx),
+				serviceOrderReview.user().associate(user, trx),
+			]);
+
+			await commit();
+		} catch (error) {
+			await trx.rollback();
+			throw error;
+		}
+
+		return serviceOrderReview.toJSON();
+	}
+
+	/**
 	 * Close TechnologyOrder.
 	 * PUT orders/:id/close
 	 */
-	async closeOrder({ params, request, response }) {
+	async closeTechnologyOrder({ params, request, response }) {
 		const data = request.only(['unit_value', 'quantity']);
 		const order = await TechnologyOrder.findOrFail(params.id);
 		if (order.status !== orderStatuses.OPEN) {
@@ -225,7 +355,7 @@ class OrderController {
 	 * Cancel TechnologyOrder.
 	 * PUT orders/:id/cancel
 	 */
-	async cancelOrder({ params, request, response, auth }) {
+	async cancelTechnologyOrder({ params, request, response, auth }) {
 		const { cancellation_reason } = request.all();
 		const order = await TechnologyOrder.findOrFail(params.id);
 		if (order.status !== orderStatuses.OPEN) {
@@ -258,6 +388,47 @@ class OrderController {
 		};
 		Bull.add(SendMailJob.key, mailData, { attempts: 3 });
 		return order.toJSON();
+	}
+
+	async performServiceOrder({ params }) {
+		const serviceOrder = await ServiceOrder.findOrFail(params.id);
+		serviceOrder.status = serviceOrderStatuses.PERFORMED;
+		await serviceOrder.save();
+		return serviceOrder;
+	}
+
+	async destroyServiceOrder({ params, request, response }) {
+		const serviceOrder = await ServiceOrder.findOrFail(params.id);
+		const result = await serviceOrder.delete();
+		if (!result) {
+			return response
+				.status(400)
+				.send(
+					errorPayload(
+						errors.RESOURCE_DELETED_ERROR,
+						request.antl('error.resource.resourceDeletedError'),
+					),
+				);
+		}
+
+		return response.status(200).send({ success: true });
+	}
+
+	async destroyServiceOrderReview({ params, request, response }) {
+		const serviceOrderReview = await ServiceOrderReview.findOrFail(params.id);
+		const result = await serviceOrderReview.delete();
+		if (!result) {
+			return response
+				.status(400)
+				.send(
+					errorPayload(
+						errors.RESOURCE_DELETED_ERROR,
+						request.antl('error.resource.resourceDeletedError'),
+					),
+				);
+		}
+
+		return response.status(200).send({ success: true });
 	}
 }
